requirements.txt
# Core libraries
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
numpy>=2.0.0
pandas>=2.0.0
scipy>=1.11.0

# Image processing
scikit-image>=0.21.0
Pillow>=10.0.0

# ML utilities
xgboost>=2.0.0
optuna>=3.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Application
streamlit>=1.28.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0